Moved the packet registration stuff out of the Packet class
